Better file organization that takes into account components
