+++ src/components/ModeIndicator/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var React = require('react');
-var store = require('../store');
+var store = require('../../store');
 
 /**
  * ## Mode Indicator Component
