--- src/helpers/optionDefaults.js
@@ -1,24 +1,0 @@
-'use strict';
-
-/**
- * # Default options that Data Lasso is initialized with
- *
- * All options that can be overwritten from outside of Data Lasso defined below.
- */
-module.exports = {
-
-    /**
-     * Size of the graph
-     */
-    graphSize: 2000,
-
-    /**
-     * Size of the axis legend text
-     */
-    legendSize: 50,
-
-    /**
-     * Color of the legend text
-     */
-    legendColor: 0xffffff,
-};
