+++ src/store/index.js
@@ -2,9 +2,9 @@
 
 var _ = require('lodash');
 var Model = require('backbone').Model;
-var datahelper = require('../helpers/data');
-var initialState = require('./initialState');
 var dispatcher = require('../dispatcher');
+var dataFunctions = require('./data-functions');
+var initialState = require('./initial-state');
 
 /**
  * # Store
@@ -91,7 +91,7 @@ var DataModel = Model.extend({
         if (!action.entries) {
             throw new Error('New data does not contain any entries');
         }
-        var data = datahelper.processInput(action.entries, this.options);
+        var data = dataFunctions.processInput(action.entries, this.options);
         this.set({
             entries: data.entries,
             attributes: data.attributes,
@@ -149,7 +149,7 @@ var DataModel = Model.extend({
         }
         this.set({
             mappings: action.mappings,
-            scales: datahelper.getUpdatedScales(this.get('entries'), this.options),
+            scales: dataFunctions.getUpdatedScales(this.get('entries'), this.options),
         });
     },
 
@@ -178,7 +178,7 @@ var DataModel = Model.extend({
         this.set({
             entries: newEntries,
             selectedEntries: [],
-            scales: datahelper.getUpdatedScales(newEntries, this.options),
+            scales: dataFunctions.getUpdatedScales(newEntries, this.options),
         });
     },
